Break TritonGPUToLLVMPass into TritonGPUToRockPass and RockToLLVMPass
This PR splits the TritonGPUToLLVMPass into two passes without
changing or adding any operators.
The purpose of this PR is to expose the point where rock operators and new
operators can be inserted.

There are 7 steps in the original
TritonGPUToLLVMPass. TritonGPUToRockPass takes the first 2 and
RockToLLVMPass takes the other 5.
